Makefile: add a modules_install target
Allows to cleanly install the module in the standard way, as suggested by
Documentation/kbuild/modules.txt in the kernel sources.

This is needed when installing from within a build system (such as Buildroot)
that implement depmod on its own, and does not need the blacklist and other
fixups that are implemented by the 'install' target.

Signed-off-by: Luca Ceresoli <luca@lucaceresoli.net>