rtl8188eu: Update README.md and Makefile
I will no longer support the download of zip versions for the source.
whenever there is any change in the source, the user must redo the
entire download and unpacking. This sequence is a lot more complicated
that a simple 'git pull', thus the user skips that step and I have to
deal with complaints about a problem I have already fixed. Now, the
make operation will fail with an error.

Signed-off-by: Larry Finger <Larry.Finger@lwfinger.net>